+++ src/test/java/lotto/utils/LottoNumberGeneratorTest.java
@@ -1,6 +1,7 @@
 package lotto.utils;
 
 import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
